requirements.txt
amqp==2.2.2
asgi-redis==1.4.3
asgiref==1.1.2
async-timeout==2.0.1
attrs==17.4.0
autobahn==17.10.1
Automat==0.6.0
billiard==3.5.0.3
celery==4.1.0
channels==1.1.8
click==6.7
constantly==15.1.0
daphne==1.4.2
dj-database-url==0.4.2
Django==2.0.1
django-choices==1.6.0
hyperlink==17.3.1
incremental==17.5.0
kombu==4.1.0
msgpack-python==0.5.1
psycopg2==2.7.3.2
python-dotenv==0.7.1
pytz==2017.3
redis==2.10.6
six==1.11.0
Twisted==17.9.0
txaio==2.8.2
vine==1.1.4
whitenoise==3.3.1
zope.interface==4.4.3
